Add files via upload
Spreadsheet version of output.txt. Gives us an approximate idea of what algorithm is doing, but unfortunately I believe there are things not being printed out by the CUDA threads. It seems like only blockIdx of 14,15 and 15,15 are printing. Each has 32^2 threads (threadIdx of 0~31,0~31), so in total we are only getting 2048 out of the 32^3 (32768, or 6.25%) potential outputs.

Anyway it gives us some insight as to how to tackle this problem. I would say the following two things are things we can proceed to experiment on:
1. Try to parallelize the for-loop in the mm_kernel() function. Maybe instead of a 2D block of CUDA threads (or is that how the architecture works?) we can try a 3D block? Not sure how it'd speed up, but then again, lots of other things for this assignment (as well as other topics) that I'm not sure about.
2. Mess around with block size and dimensions size? Again, no idea how it'd help.

I'll copy and paste this tl;dr in the comment to this file again, so as to fac